Remove dependency on BouncyCastle library.
This change removes the dependency on BouncyCastle in the Multipaz
libraries, except for `multipaz-android-legacy` (more at the end about
this).

The root problem is that `CryptoJvm`'s initializer did the following

```
    Security.removeProvider(BouncyCastleProvider.PROVIDER_NAME)
    Security.insertProviderAt(BouncyCastleProvider(), 1)
```

which is fundamentally wrong as a library should never modify
process-wide state behind the user's back. Remove this.

The other problem is that we were using various functions/interfaces
from BouncyCastle library that we didn't really need to use anymore,
especially since we implemented our own code for parsing/generating
ASN.1 and X.509 certificates.

After this, another problem remains. Since the application is now in
charge of managing which JCA providers to use - as it should - we need
to handle the situation where the Java `PublicKey` and `PrivateKey`
instances we're dealing with are from arbitrary providers
(e.g. `AndroidOpenSSL` which is the name of the Conscrypt provider
which is the default on Android) instead of always assuming they're
from BouncyCastle.

In samples/testapp, add a new setting "Prefer BouncyCastle to
Conscrypt" toggle which defaults to off, meaning that by default our
TestApp will use Conscrypt instead of BouncyCastle. This does mean
that Brainpool curves won't work (as Conscrypt doesn't currently
support that) but everything else works and of course Brainpool curves
will work when this new setting is on. Also restart the app when this
is toggled and use ProcessPhoenix for this since is in fact
non-trivial.

Adjust `Crypto.supportedCurves` to not include Brainpool curves unless
BouncyCastle is first in the list. Also add `Crypto.provider` string
which includes the list of JCA providers when using `CryptoJvm`. Print
this string in _Software Secure Area_ screen in the testapp.

For testing, use BouncyCastle when running tests on JVM and not when
running on Android. This helps ensure `CryptoJvm` gets tested with
different sets of JCA providers.

For `multipaz-android-legacy` remove all calls to
`BountyCastleProvider()` and adjust the tests so they always run
against the default `AndroidOpenSSL` provider on Android
(Conscrypt). Still link against BouncyCastle library because it's
being using to generate certificates and other things. While we could
port this to using the ASN1 and X.509 routines in the core Multipaz
library such a change adds risk and we'd rather not make too big
changes. Also, `multipaz-android-legacy` will be removed in the near
future as known downstreams are actively migrating away from it.

Carefully tested this with TestApp going through the _Software Secure
Area_, _Cloud Secure Area_, _Android Secure Area_, and _Secure Enclave
Secure Area_ screens testing all combinations. On Android this was
done with the new "Prefer BouncyCastle to Conscrypt" setting in both
possible places.

Also remove BouncyCastleProvider() calls in `wallet` and `appverifier`
modules and test those.

As a result, the only dependencies on BouncyCastle we now have are for
testing the core library, in `multipaz-android-legacy`, in
`multipazctl` command-line tool (to generate keys/certs with Brainpool
curves), and for running the Cloud Secure Area server-side (so it can
support Brainpool curves). This should be good news for downstream
consumers on Android as this allows them to not include BouncyCastle
at all (unless they need Brainpool curve support) which simplifies
supply chain management and also helps reduce APK size.

Also rename package and file for `multipazctl` as that seems to have
been forgotten in the big rename we had a while ago.

Test: New unit tests and all unit tests pass.
Test: Tested `samples/testapp` manually on Android and iOS (see above).
Test: Tested `wallet` and `appverifier`.
Signed-off-by: David Zeuthen <zeuthen@google.com>