Consent prompt updates and export Credman/NFC services/actvities in library. (#1115)
This mainly consists of visual updates to ConsentModalBottomSheet to
make it look more crisp. This includes adding `privacyPolicyUrl` to
`TrustPoint` and showing that as a link in the consent dialog.

Also change the flow in the Credman presentment so the activity is on
translucent and on top of the browser/app invoking it and don't show
the "The data was shared" or "Something went wrong" animations.

Also export testapp's code for Credman and NFC engagement and
presentment in reusable classes in multipaz-compose, see the
latest commit in https://github.com/davidz25/MpzCmpWallet for
an example.

This change also renames the package name from com.android.identity.testapp
to org.multipaz.testapp which we forgot to do when we originally
renamed Multipaz.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>