Implement SecureArea.batchCreateKey() for Multipaz Cloud Secure Area.
This primary motivation for this was Issue #1125 in order to assess if
optimizations for Multipaz are needed.

This change also implements OpenID4VCI key attestations for
`batchCreateKey()` and implementing this made me realize that we used
the wrong return type for these attestations. We were returning
`JsonObject` for the body of the attestation but in order for this to
be useful the caller really needs the whole JWS, including header and,
crucially, the signature. So change this to return a String and
specify it should be the compact serialization encoding of the
JWS. Existing users can easily migrate, it didn't really work anyway
before this.

In DocumentStoreScreen in the Testapp also add two new buttons "Create
EUPID in CSA w/ 10 creds" and "Create EUPID in CSA w/ 10 creds
(batch)" to easily test batching and compare timings against
non-batching. Also show show a dialog when provisioning is done. This
shows the amount of time it took and also the OpenID4VCI attestation
in case one is returned.

For Multipaz Cloud Secure Area it takes ~2 seconds on Android and ~1
second on iOS to create a PID with 10 credentials using batching when
using a CSA with 120ms pings (Berlin to US Central). This seems to be
acceptable and on Android we can even get a bit faster, see the
ongoing discussion in #1125.

Also change the settings for Cloud Secure Area to use snake_case
instead of CamelCase. Add new settings to configure OpenID4VCI
attestation, in particular `iss`, `certification`, `key_storage`, and
`authentication` values in the returned JWT.

Also rearrange things in the TestApp so it's using the
org.multipaz.util.Platform helpers instead of its own
abstractions. This makes it easier to wholesale e.g. use
EphemeralStorage() in a single place which is handy when evaluating
performance.

Test: New unit tests and all unit tests pass.
Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>