Rework TrustManager. (#1135)
This changes `TrustManager` to be an interface and introduces
`LocalTrustManager` which is backed by local files,
`VicalTrustManager` backed by a VICAL according to ISO/IEC 18013-5,
and `CompositeTrustManager` which can be used to stack trust managers.

The idea is that `LocalTrustManager` can be used for storing to those
trust points (verifiers or issuers) that the user explicitly chooses
to trust and its API allow for adding/remove trust points.

It also changes `TrustPoint` to be a sealed class with concrete
classes `X509CertTrustPoint` and `OriginTrustPoint`.

Also add two new screens in Test App for showing trusted issuers and
trusted readers.

Disable use of TrustManager in `wallet` module since a lot of that
codebase is assuming non-suspend functions. This module will be
deleted soon so this functionality (lack of reader auth) isn't a major
loss.

Future commits will bring `RicalTrustManager` (when we add RICAL
support) and ApplicationTrustPoint (for same-device apps requesting
credentials).

Test: New unit tests and all unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>