QrCodeScanner: Fix problem with onCodeScanned() being called multiple times. (#1139)
This problem manifested itself in the _ISO mdoc Proximity Reading_
screen insofar that `doReaderFlow()` occasionally would be invoked
multiple times leading to the wrong SessionTranscript to be used
because a new ephemeral reader key was generated. This resulted in a
"Device Authentication failed" warning when showing the result.

This problem was introduced in Commit fe28caf where we changed the QR
code scanning library on Android to be ZXing.

Test: Manually tested and observed "Device Authentication failed" no longer is shown.

Signed-off-by: David Zeuthen <zeuthen@google.com>