scanNfcDialog: Allow NFC scanning without showing dialog.
On Android, it's possible to do NFC scanning without any UI and this
is desirable in e.g. an mdoc identity reader app. Add support by this
by allowing the passed-in message to be null. Also add a flag so
applications can test at runtime whether this functionality is
available.

Also include a new reader trust point in samples/testapp.

Test: New test items in samples/testapp.
Test: Manually tested on both Android and iOS.
Signed-off-by: David Zeuthen <zeuthen@google.com>