First cut of system-of-record server.
System-of-record allows admins to create new test identities and specify core data and additional
records associated with these identities. It is then accessed from the provisioning server
(openid4vci) as source of data for the credentials.

This first cut is done in simplistic way; in particular, there is not yet a way to authenticate
admins (anyone can add/edit/remove identities). Also, the integration between the system-of-record
and openid4vci servers is very simplistic, in the future, system-of-record will serve as
authorization server for openid4vci instead.

Example of openid4vci server invocation to use a system-of-record server as back-end:
```
./gradlew multipaz-openid4vci-server:run -- --args="-param system_of_record_url=http://localhost:8004"
```

Test: manually tested with the existing (tweaked in this PR) openid4vci server.

Signed-off-by: Peter Sorotokin <sorotokin@gmail.com>