TrustManager and CertificateViewer changes.
For TrustManager, drop the functionality to authenticate by origin
since this isn't a realiable mechanism and doesn't work well when
using different-device flow using e.g. CTAP/Hybrid. Make it all X.509
certificate based which simplifies everything.

Also rename LocalTrustManager to TrustManagerLocal and introduce a
richer API for managing entries that contain trust points. This
includes both X.509 certificate roots but also VICALs. Also add
functionality to update entries with new names / icons.

This API now lends itself nicely to what a reader or wallet app would
include in its Settings screen to manage these entries, see
https://github.com/davidz25/MpzIdentityReader for an example of an app
using this API.

Also rename CertificateViewer to X509CertViewer, remove the built-in
scrolling and remove the variant that takes a X509CertChain since apps
will want to do this themselves for non-trivial embeddings. The
problem with the original code was that it wasn't really possible to
embed on a screen with other graphical elements because it did its own
scrolling. Also update all text labels in X509CertViewer to be
compliant with the Material spec which calls for not capitalizing each
word in a sentence.

Test: New unit tests and all unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>