OAuth-based integration between openid4vci server and the sytem of record. (#1153)
When running with the system of record, openid4vci server now delegates
authorization to the system of record itself. Data is now pulled from the system
of record at the time of credential issuance. Also, system of record
access from openid4vci server is secured in the typical oauth fashion: only
trusted servers are allowed to connect. (Admin interface is still not
authenticated, that will come later).

Test: tested using testapp both with and without using the system of record.

Signed-off-by: Peter Sorotokin <sorotokin@gmail.com>