Add support for presentment via URI schemes. (#1164)
Obviously W3C Digital Credentials API is preferable but it looks like
URI schemes will see some use in the short- to mid-term.

This PR adds support for URI schemes for OpenID4VP 1.0 and it includes
support for both the verifier and the wallet side. This also includes
a slight cleanup to the verifier codebase and adds support for using
TrustManager to determine trust points. Initially this contains the
IACA cert for testapp credentials but in the future we can add more.

Additionally, this PR moves all the OpenID4VP presentment code to the
multipaz-models library.

Test: Manually tested

Signed-off-by: David Zeuthen <zeuthen@google.com>