Add support for ZKP over the Internet. (#1169)
This adds support in our verifier and wallet code for both 18013-7
Annex C and OpenID4VP. Because of temporary limitations in Credman
this only works with URI schemes because of the size of the ZKP
response. I'm told this will be fixed in a future Credman release.
Concretely this means the only way to test ZKP over the Internet right
now is to use "URI scheme (OpenID4VP 1.0)" from the protocol dropdown
on the verifier.

Other changes:

- Make testapp build/run on iOS, the previous commit related to ZKP broke that.
- Make sure we're including namespaces in ZkDocumentData according to latest
  18013-5 Second Edition draft.
- Minor cleanups / code organization changes.

Test: Manually tested with verifier running on https://ws.davidz25.net
Test: Manually tested proximity presentment with ZKP.

Signed-off-by: David Zeuthen <zeuthen@google.com>