Digital Credentials: Support multi-protocol requests.
Introduce the following new options for the Protocol menu in the Online Verifier:

- Annex C + OpenID4VP 1.0
- Annex C + OpenID4VP Draft 24
- OpenID4VP 1.0 + Annex C
- OpenID4VP Draft 24 + Annex C

where each new combination is a W3C Digital Credential request with
two elements in the `requests` field. It is expected that RPs in some
ecosystems will have to make requests like these in order to support
both Google Wallet (OpenID4VP) and Apple Wallet (Annex C).

Update our matcher and Multipaz to properly support this, in
particular make it so we honor the request in order they appear. In
other words for "Annex C + OpenID4VP 1.0" return a credential using
the `org-iso-mdoc` exchange protocol, for "OpenID4VP 1.0 + Annex C"
return one for the `openid4vp-v1-signed` exchange protocol.

Also add API to `DigitalCredentials` for the application to configure
which W3C DC protocols it wants to support. By default all supported
protocols are enabled but some wallets might want to limit this to
e.g. only OpenID4VP 1.0 or ISO/IEC 18013-7:2025 Annex C or
whatever. Also expose this in samples/testapp on the Settings page
since it's helpful for testing fallbacks for RPs.

Also add a couple of more examples for Raw DCQL. This is in
anticipation of implementing better DCQL support in both the matcher
and library in the near future.

Test: Manually tested.
Test: All unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>