DeviceResponseParser: Support out-of-spec responses. (#1183)
A wallet vendor makes available test mDLs but the responses for these
are unfortunately not compliant with ISO/IEC 18013-5:2021 because they
lack an issuer signature. We can however easily support these since we
already have an `issuerSignedAuthenticated` signal that we can just
set to `false`. Also update online verifier to cope with the `X509CertChain`
in `issuerCertificateChain` property having zero certificates.

Test: New unit test and all unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>