Add tests for Credman matcher.
This adds a new Android instrumented tests which links with a C++
library that includes the matcher but where the Credman-specific stubs
are replaced by stubs that interact with our testing
harness. Consequently we can feed the matcher a W3C Digital
Credentials array of requests and we get back a list of credentials
that it added to the picker, including the fields that were shown to
the user. This makes it possible to check that the Credential Manager
UI is showning exactly what we expect it to show.

Test: New unit test and all unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>