Migrate to kotlinx-datetime 0.7.1.
The [migration to kotlix-datetime 0.7.1](https://github.com/Kotlin/kotlinx-datetime?tab=readme-ov-file#deprecation-of-instant) is slightly complicated
because `Clock` and `Instant` moved from the `kotlinx.datetime`
package to the `kotlin.time` package.

Also update versions of other most dependencies, removed unused
dependencies, and fix some compiler warnings.

Also include multipaz-longfellow in our XCFramework and publish a
snapshot of this so we can test it works before the 0.93 release.

Test: Unit tests.
Test: Manually tested on both Android and iOS.
Signed-off-by: David Zeuthen <zeuthen@google.com>