Add support for conveying Google Account in reader auth cert.
Concretely, add a `googleAccount` map entry to the top-level map in
the CBOR for our Multipaz X.509 extension which can contain
information about the signed-in Google Account in the reader
application making the request.

Also add a `dynamicMetadataResolver` argument to both `PresentmentSource`
interface and the `CredentialPresentmentModalBottomSheet` composable
so an application can dynamically provide `TrustMetadata` based
on e.g. data in the leaf reader certificate. Also add new fields
`displayIconUrl` and `disclaimer` to `TrustMetadata` and support for
loading icons over the network using [Coil](https://coil-kt.github.io/coil/).

In TestApp, use this new infrastructure to check for `googleAccount`
on reader cerficates which are signed by the Multipaz Identity Reader
certificates and in a separate commit modify the reader to set this
information.

As a result, testapp will now show the Google account email and
profile picture when using a modified version of Multipaz Identity
Reader which supports this.

Also update the testapp screen for Consent Page with comboboxes
instead of a long list and add support for this new Google Account
feature and also more complicated multi-document requests.

Test: New unit tests and all unit tests pass.
Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>